refactor: modularize JS, CSS, and add CSS mask-image icons
JavaScript:
- Split content-ideas.js (1012 lines) into 7 modular files:
  - content-strategy-utils.js: Shared utilities and AJAX handler
  - content-strategy-generate.js: Generate button behaviors
  - content-strategy-editable.js: Editable field auto-save
  - content-strategy-delete.js: Card and idea deletion
  - content-strategy-checkbox.js: Implemented checkbox toggle
  - content-strategy-links.js: Link add/edit functionality
  - content-strategy-export.js: CSV export
- Each module uses Drupal behaviors pattern for independent attachment

CSS:
- Split content-strategy.css into 7 component files:
  - base.css: Container, status area, actions, empty state
  - cards.css: Card layout, header, priority badges
  - table.css: Content ideas table, implemented row styling
  - checkbox.css: Custom checkbox with checkmark animation
  - links.css: Link display, edit button, input form
  - delete.css: Delete buttons for cards and ideas
  - editable.css: Editable fields with save indicators
- Added icons.css with CSS mask-image for colorable SVG icons

Icons:
- Replaced inline SVG with CSS mask-image approach
- Icons now inherit color from parent via currentColor
- Updated templates to use cs-icon classes
- Simplified module preprocess (no longer passes icons via JS)

Constants:
- Added src/Constants.php with centralized magic strings
- Includes CSS selectors, field names, data attributes, icon names
