+++ templates/ai-content-strategy-recommendations-items.html.twig
@@ -13,21 +13,6 @@
  */
 #}
 
-{# Trash icon SVG #}
-{% set trash_icon %}
-<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor" aria-hidden="true">
-  <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6z"/>
-  <path fill-rule="evenodd" d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z"/>
-</svg>
-{% endset %}
-
-{# Edit icon SVG #}
-{% set edit_icon %}
-<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor" aria-hidden="true">
-  <path d="M13.488 2.513a1.75 1.75 0 0 0-2.475 0L3.305 10.22a.75.75 0 0 0-.2.336l-.828 2.899a.75.75 0 0 0 .927.927l2.899-.828a.75.75 0 0 0 .336-.2l7.707-7.707a1.75 1.75 0 0 0 0-2.475l-.658-.659ZM12.073 3.573l.659.659a.25.25 0 0 1 0 .354l-7.707 7.707-1.586.453.453-1.586 7.707-7.707a.25.25 0 0 1 .354 0l.12.12Z"/>
-</svg>
-{% endset %}
-
 {% for item in items %}
   {# Determine title field - try common fields in order #}
   {% set item_title = item.title|default(item.topic|default(item.content_type|default(item.signal|default('Recommendation'|t)))) %}
@@ -42,7 +27,7 @@
       <div class="recommendation-header">
         <h4 contenteditable="true" data-field="title" data-original="{{ item_title }}" class="editable-field">{{ item_title }}</h4>
         <button type="button" class="delete-card-link" data-section="{{ section }}" data-title="{{ item_title }}" title="{{ 'Delete this recommendation (cannot be undone)'|t }}" aria-label="{{ 'Delete recommendation: @title'|t({'@title': item_title}) }}">
-          {{ trash_icon }}
+          <span class="cs-icon cs-icon--trash" aria-hidden="true"></span>
         </button>
       </div>
       {# Try to find description - try common fields in order #}
@@ -76,7 +61,7 @@
                     {% if idea_link %}
                       <a href="{{ idea_link }}" target="_blank" class="idea-link" data-section="{{ section }}" data-title="{{ item_title }}" data-idea-index="{{ key }}">{{ idea_link }}</a>
                       <button type="button" class="idea-link-edit" data-section="{{ section }}" data-title="{{ item_title }}" data-idea-index="{{ key }}" title="{{ 'Edit link'|t }}">
-                        {{ edit_icon }}
+                        <span class="cs-icon cs-icon--edit cs-icon--sm" aria-hidden="true"></span>
                       </button>
                     {% else %}
                       <button type="button" class="idea-add-link action-link" data-section="{{ section }}" data-title="{{ item_title }}" data-idea-index="{{ key }}">{{ '+ Add link'|t }}</button>
@@ -89,7 +74,7 @@
                     <span class="idea-checkbox-visual"></span>
                   </label>
                   <button type="button" class="delete-idea-link" data-section="{{ section }}" data-title="{{ item_title }}" data-idea-index="{{ key }}" title="{{ 'Delete this content idea'|t }}" aria-label="{{ 'Delete content idea'|t }}">
-                    {{ trash_icon }}
+                    <span class="cs-icon cs-icon--trash cs-icon--md" aria-hidden="true"></span>
                   </button>
                 </td>
               </tr>
