+++ templates/components/link-display.html.twig
@@ -14,7 +14,5 @@
 #}
 <a href="{{ link }}" target="_blank" class="idea-link" data-section="{{ section }}" data-title="{{ title }}" data-idea-index="{{ idea_index }}">{{ link }}</a>
 <button type="button" class="idea-link-edit" data-section="{{ section }}" data-title="{{ title }}" data-idea-index="{{ idea_index }}" title="{{ 'Edit link'|t }}">
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor" aria-hidden="true">
-    <path d="M13.488 2.513a1.75 1.75 0 0 0-2.475 0L3.305 10.22a.75.75 0 0 0-.2.336l-.828 2.899a.75.75 0 0 0 .927.927l2.899-.828a.75.75 0 0 0 .336-.2l7.707-7.707a1.75 1.75 0 0 0 0-2.475l-.658-.659ZM12.073 3.573l.659.659a.25.25 0 0 1 0 .354l-7.707 7.707-1.586.453.453-1.586 7.707-7.707a.25.25 0 0 1 .354 0l.12.12Z"/>
-  </svg>
+  <span class="cs-icon cs-icon--edit cs-icon--sm" aria-hidden="true"></span>
 </button>
